+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/ui/content/ContentPanelHandler.kt
@@ -67,6 +67,10 @@ internal open class ContentPanelHandler(
     }
   }
 
+  fun showTool(id: String) {
+    toolsMenuTree.selectDeveloperTool(id) {}
+  }
+
   protected open fun createDeveloperToolContentPanel(developerToolNode: DeveloperToolNode): DeveloperToolContentPanel =
     DeveloperToolContentPanel(developerToolNode)