+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/ui/instance/toolwindow/MainToolWindowService.kt
@@ -18,7 +18,7 @@ internal class MainToolWindowService(val project: Project) {
 
   private val toolWindowLock = ReentrantReadWriteLock()
   private var toolWindow: ToolWindow? = null
-  private var deferredOpenToolTask: OpenToolTask<*>? = null // Also only used under `toolWindowLock`
+  private var deferredToolTask: ToolTask? = null // Also only used under `toolWindowLock`
 
   // -- Initialization ---------------------------------------------------------------------------------------------- //
   // -- Exported Methods -------------------------------------------------------------------------------------------- //
@@ -26,9 +26,12 @@ internal class MainToolWindowService(val project: Project) {
   fun setToolWindow(toolWindow: ToolWindow) {
     toolWindowLock.write {
       this.toolWindow = toolWindow
-      if (deferredOpenToolTask != null) {
-        doOpenTool(toolWindow, deferredOpenToolTask!!.context, deferredOpenToolTask!!.reference)
-        deferredOpenToolTask = null
+      if (deferredToolTask != null) {
+        when (val toolTask = deferredToolTask!!) {
+          is OpenToolTask<*> -> doOpenTool(toolWindow, toolTask.context, toolTask.reference)
+          is ShowToolTask -> doShowTool(toolWindow, toolTask.id)
+        }
+        deferredToolTask = null
       }
     }
   }
@@ -40,11 +43,24 @@ internal class MainToolWindowService(val project: Project) {
         doOpenTool(toolWindow!!, context, reference)
       }
       else {
-        deferredOpenToolTask = OpenToolTask(context, reference)
+        deferredToolTask = OpenToolTask(context, reference)
         // The content of the tool window will be added asynchronous. So, the
         // return of `show()` (or its callback equivalents) will not indicate
         // the `toolWindowContentPanelHandlerKey` was already set. Therefore,
         // we have to defer the open tool task.
+        ToolWindowManager.getInstance(project).getToolWindow(MainToolWindowFactory.ID)?.show()
+      }
+    }
+  }
+
+  fun showTool(id: String) {
+    toolWindowLock.read {
+      if (toolWindow != null) {
+        toolWindow!!.show()
+        doShowTool(toolWindow!!, id)
+      }
+      else {
+        deferredToolTask = ShowToolTask(id)
         ToolWindowManager.getInstance(project).getToolWindow(MainToolWindowFactory.ID)?.show()
       }
     }
@@ -62,12 +78,29 @@ internal class MainToolWindowService(val project: Project) {
       ?.openTool(context, reference)
   }
 
+  private fun doShowTool(
+    toolWindow: ToolWindow,
+    id: String
+  ) {
+    toolWindow.contentManager.getContent(0)
+      ?.getUserData(toolWindowContentPanelHandlerKey)
+      ?.showTool(id)
+  }
+
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 
-  private data class OpenToolTask<T : OpenDeveloperToolContext>(
+  private sealed interface ToolTask
+
+  // -- Inner Type -------------------------------------------------------------------------------------------------- //
+
+  private class OpenToolTask<T : OpenDeveloperToolContext>(
     val context: T,
     val reference: OpenDeveloperToolReference<out T>
-  )
+  ) : ToolTask
+
+  // -- Inner Type -------------------------------------------------------------------------------------------------- //
+
+  private class ShowToolTask(val id: String) : ToolTask
 
   // -- Companion Object -------------------------------------------------------------------------------------------- //